requirements.txt
beartype==0.22.8
fastapi[all]
uvicorn[standard]
slowapi
python-multipart
python-dotenv==1.2.1
PyYAML==6.0.2
shortuuid==1.0.13
jinja2
websockets
numpy==1.23.5
opencv-python==4.10.0.84
opencv-python-headless==4.6.0.66
PyJWT==2.10.1
pyserial==3.5
cachetools==5.3.3